Move remaining addon test buttons into their own windows
